refactor: Update settleBet function to settle wagers for SOL and MUSHU tokens
This commit refactors the settleBet function in the bet module to include the settlement of wagers for SOL and MUSHU tokens. It replaces the commented out code that was previously used to settle SOL wagers and adds similar functionality for MUSHU wagers. The console.info statements have been updated to log the length of the wagers array instead of the user addresses. The settleBets function in the bet usecase module has also been updated to order the bets by descending startedAt value.